server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiohttp>=3.10.0
python-chess>=1.999
pydantic>=2.0.0
